Use an HTTP client to get result from the HAL website
Javalin runs on top of Jetty. So I can use classes and fonctions
from it. I think this is a good thing for people who used to work with Jetty
(I did not use Jetty before but whatever...)

However, Javalin does not provide a high-level HTTP client on top of Jetty that support JSON (de)serialization.
-> Spring provides something like that with RestTemplate
Even if the Jetty's HttpCient is not complicated to use (4 lines of code to get result from an external),
A dedicated HTTP client provided by Javalin with JSON serialization/deserialization would be great.

Maybe the author wanted to keep Javalin as small and lightweight as possible.

I will write an feature request just in case.